implement initial universe container with cells
