+++ jpcd-yew/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "trunk-template"
+name = "jp-conway"
 version = "0.1.0"
 edition = "2021"
 description = "Template for starting a Yew project using Trunk"
@@ -12,3 +12,7 @@ categories = ["gui", "wasm", "web-programming"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
 yew = { version="0.20", features=["csr"] }
+log = "0.4"
+rand = "0.8"
+wasm-logger = "0.2"
+gloo = "0.8"
